# docs/swagger-openapi.yml
openapi: 3.0.1
info:
  title: Tài liệu REST API Nexora-Commerce
  description: Tài liệu RESTful API hiệu năng cao cho hệ thống thương mại điện tử
    Nexora-Commerce phục vụ tích hợp frontend.
  contact:
    name: Nexora Core Dev Team
    email: dev@nexoracommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Máy chủ phát triển cục bộ (Local Development Server)
security:
- bearerAuth: []
...
paths:
...
  /api/v1/admin/orders/{orderId}/status:
    put:
      tags:
      - Quản lý Đơn hàng (Admin)
      summary: Cập nhật trạng thái đơn hàng (Admin)
      description: Yêu cầu quyền ADMIN. Cập nhật trạng thái xử lý chính thức của một
        đơn hàng.
      operationId: updateOrderStatus
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
      - name: status
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
      security:
      - bearerAuth: []
...
  /api/v1/orders:
    get:
      tags:
      - Đơn hàng Người dùng
      summary: Lấy lịch sử đơn hàng của khách hàng
      description: Lấy danh sách lịch sử đơn hàng đã đặt phân trang của người dùng
        cụ thể. Yêu cầu ID người dùng khớp hoặc quyền ADMIN.
      operationId: getAllOrders_1
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      - name: userId
        in: query
        description: ID khách hàng liên kết với các đơn hàng
        required: true
        schema:
          type: string
        example: USR-001
      responses:
        '403':
          description: Không đủ quyền truy cập
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListOrderResponse'
        '200':
          description: Lấy danh sách trang đơn hàng của khách hàng thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListOrderResponse'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListOrderResponse'
      security:
      - bearerAuth: []
    post:
      tags:
      - Đơn hàng Người dùng
      summary: Tạo đơn hàng đặt trước
      description: Xử lý chi tiết giỏ hàng hiện tại và tạo một giao dịch đặt hàng
        chờ thanh toán. Yêu cầu ID người dùng khớp hoặc quyền ADMIN.
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
        required: true
      responses:
        '201':
          description: Tạo đơn hàng thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
        '403':
          description: Không đủ quyền truy cập
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
        '400':
          description: Tham số không hợp lệ hoặc giỏ hàng trống
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
      security:
      - bearerAuth: []
...
  /api/v1/orders/{orderId}/status:
    patch:
      tags:
      - Đơn hàng Người dùng
      summary: Cập nhật trạng thái đơn hàng (Admin)
      description: 'Yêu cầu quyền ADMIN. Cập nhật và chuyển đổi trạng thái của đơn
        hàng (ví dụ: PENDING, PROCESSING, COMPLETED, CANCELLED).'
      operationId: updateOrderStatus_1
      parameters:
      - name: orderId
        in: path
        description: ID đơn hàng cần cập nhật
        required: true
        schema:
          type: string
        example: ORD-20260523-999
      - name: status
        in: query
        description: Trạng thái chuyển đổi đích của đơn hàng
        required: true
        schema:
          type: string
        example: COMPLETED
      responses:
        '200':
          description: Cập nhật trạng thái đơn hàng thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
        '403':
          description: Không đủ quyền hạn (yêu cầu vai trò ADMIN)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
        '404':
          description: Không tìm thấy ID đơn hàng
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
        '400':
          description: Chuyển đổi trạng thái không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
        '401':
          description: Mã xác thực JWT thiếu hoặc không hợp lệ
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseOrderResponse'
      security:
      - bearerAuth: []
...
  /api/v1/admin/orders:
    get:
      tags:
      - Quản lý Đơn hàng (Admin)
      summary: Lấy toàn bộ đơn hàng (Admin)
      description: Yêu cầu quyền ADMIN. Lấy danh sách đơn hàng toàn hệ thống dạng
        phân trang phục vụ quản lý đơn hàng.
      operationId: getAllOrders
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Lấy danh sách đơn hàng thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListOrderResponse'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    ApiResponseOrderResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/OrderResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        variantSku:
          type: string
        productName:
          type: string
        variantName:
          type: string
        quantity:
          type: integer
          format: int32
        price:
          type: integer
          format: int64
    OrderResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        subtotal:
          type: integer
          format: int64
        discountAmount:
          type: integer
          format: int64
        shippingFee:
          type: integer
          format: int64
        totalPrice:
          type: integer
          format: int64
        couponCode:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - CONFIRMED
          - PROCESSING
          - SHIPPED
          - DELIVERED
          - CANCELLED
        paymentStatus:
          type: string
          enum:
          - UNPAID
          - PAID
          - REFUNDED
        customerNote:
          type: string
        createdAt:
          type: string
          format: date-time
        lastModifiedDate:
          type: string
          format: date-time
        shippingAddress:
          type: string
        phoneNumber:
          type: string
        paymentUrl:
          type: string
        paymentTransactions:
          type: array
          items:
            $ref: '#/components/schemas/PaymentTransactionResponse'
    PaymentTransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
        paymentMethod:
          type: string
          enum:
          - COD
          - VNPAY
        providerTransactionId:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - SUCCESS
          - FAILED
        createdAt:
          type: string
          format: date-time
    OrderItemRequest:
      required:
      - price
      - productName
      - quantity
      - variantSku
      type: object
      properties:
        variantSku:
          type: string
        productName:
          maxLength: 255
          minLength: 0
          type: string
        variantName:
          maxLength: 255
          minLength: 0
          type: string
        quantity:
          type: integer
          format: int32
        price:
          type: integer
          format: int64
    OrderRequest:
      required:
      - orderItems
      - paymentMethod
      - phoneNumber
      - shippingAddress
      - userId
      type: object
      properties:
        userId:
          type: string
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        couponCode:
          maxLength: 50
          minLength: 0
          type: string
        shippingAddress:
          type: string
        phoneNumber:
          type: string
        customerNote:
          maxLength: 500
          minLength: 0
          type: string
        paymentMethod:
          type: string
          enum:
          - COD
          - VNPAY
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    ApiResponseListOrderResponse:
      type: object
      properties:
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Please paste your JWT Access Token in the box below to authorize.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT